新增spring和mybatis,druid整合demo

+++ thread-core/webelement/src/test/util/SpeLUtil.java
@@ -1,5 +1,6 @@
 package test.util;
 
+import org.junit.Test;
 import org.springframework.expression.Expression;
 import org.springframework.expression.ExpressionParser;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
@@ -12,7 +13,7 @@ public class SpeLUtil {
     public void testEle(){
         ExpressionParser parser = new SpelExpressionParser();
         Expression expression = parser.parseExpression("6+2");
-        Integer integer = expression.getValue();
+        Integer integer = (Integer) expression.getValue();
         System.out.println(integer);
     }
 }
